stripe moved to libraries
moved stripe to libraries and code corrections, removed unwanted things
