+++ app/controllers/ProductController.php
@@ -1,17 +1,14 @@
 <?php
 namespace App\Controllers;
 
-use App\Config\Connection;
 use App\Models\ProductModel;
 
 class ProductController
 {
-    private $pdo;
     private $productModel;
     
     public function __construct()
     {
-        $this->pdo = Connection::make();
         $this->productModel = new ProductModel();
     }
     
@@ -19,5 +16,17 @@ class ProductController
     {
         $products = $this->productModel->getAllProducts();
         return $products;
+    }
+
+    public function validatingProduct($productId)
+    {
+        $condition = ["product_id" => $productId];
+        $productData =  $this->productModel->getProductData($condition); 
+        if (empty($productData)) {
+            return ["status" => "0", "message" => "Product is not available !"];  
+        }
+        if ($productData->product_stock <= 0) { 
+            return ["status" => "0", "message" => "Product is out of stock"];
+        } 
     }
 }
